+++ packages/legend-query-builder/src/__lib__/QueryBuilderTesting.ts
@@ -27,6 +27,7 @@ export enum QUERY_BUILDER_TEST_ID {
   QUERY_BUILDER_EXPLORER = 'query__builder__explorer',
   QUERY_BUILDER_PROPERTY_SEARCH_PANEL = 'query__builder__property__search__panel',
   QUERY_BUILDER_RESULT_PANEL = 'query__builder__result__panel',
+  QUERY_BUILDER_RESULT_ANALYTICS = 'query__builder__result__analytics',
   QUERY_BUILDER_PARAMETERS = 'query-builder__parameters',
   QUERY_BUILDER_CONSTANTS = 'query-builder__constants',
 }