+++ packages/legend-application-query-bootstrap/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-query-bootstrap
 
+## 10.29.0
+
+## 10.28.0
+
 ## 10.27.0
 
 ## 10.26.0