+++ packages/legend-application-query/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-query
 
+## 13.1.3
+
+## 13.1.2
+
 ## 13.1.1
 
 ## 13.1.0