+++ packages/legend-application-studio-deployment/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-studio-deployment
 
+## 10.29.0
+
+## 10.28.0
+
 ## 10.27.0
 
 ## 10.26.0