+++ packages/legend-application-studio/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-studio
 
+## 28.4.3
+
+## 28.4.2
+
 ## 28.4.1
 
 ### Patch Changes