+++ packages/legend-extension-dsl-data-space/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-data-space
 
+## 10.1.22
+
+## 10.1.21
+
 ## 10.1.20
 
 ## 10.1.19