+++ packages/legend-extension-dsl-diagram/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-diagram
 
+## 8.0.78
+
+## 8.0.77
+
 ## 8.0.76
 
 ## 8.0.75