+++ packages/legend-extension-dsl-mastery/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-mastery
 
+## 3.0.79
+
+## 3.0.78
+
 ## 3.0.77
 
 ## 3.0.76