+++ packages/legend-extension-dsl-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-text
 
+## 6.0.78
+
+## 6.0.77
+
 ## 6.0.76
 
 ## 6.0.75