+++ packages/legend-extension-store-service-store/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-store-service-store
 
+## 2.0.78
+
+## 2.0.77
+
 ## 2.0.76
 
 ## 2.0.75